Render modals in center of screen by default (PP-2243) (#54)
## Description

<!--- Describe your changes -->
Updated custom Modal component to render in the middle of the screen.
Before, the modal was rendering at the bottom of the page. This might
have been an unintended change after updating to ariakit from reakit.

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
Separating the modal so far from the event origin was bad UX.
<!--- If it fixes an open issue, please link to the issue here. -->

[Jira PP-2243](https://ebce-lyrasis.atlassian.net/browse/PP-2243)

## How Has This Been Tested?

Updated the sign out modal test to expect the new styling changes
whenever the modal is visible.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have updated the documentation accordingly.
- [x] All new and existing tests passed.