Update dependencies (#72)
## Description

Updates a bunch of dependencies and gets rid of the `@jsx` and
`jsxRuntime` pragma.

## Motivation and Context

While working on PP-3113, I was running into some issue with the older
version of `TypeScript`. This led to a cascade of dependency changes and
other clean up, but I think the code base is better for it.

## How Has This Been Tested?

- Linting, type checking, and tests pass locally.
- Built and ran locally.
- [CI test and
build](https://github.com/ThePalaceProject/web-patron/actions/runs/18857968376)
pass.

## Checklist:

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.