Display library list in ascending title order (PP-3349) (#80)
## Description

Sorts the library list in ascending order by "effective title" (the
title, if present, otherwise the slug) on the home page.

## Motivation and Context

It is more challenging to locate libraries on the home page if they are
not in sorted order.

[Jira PP-3349]

## How Has This Been Tested?

- Added new tests for the functionality.
- All tests pass locally.
- [CI
tests](https://github.com/ThePalaceProject/web-patron/actions/runs/19884045939)
pass.

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- N/A - I have updated the documentation accordingly.
- [x] All new and existing tests passed.